Introduce performance benchmark
